+++ src/units/energy.ts
@@ -1,10 +1,8 @@
 import * as UnitTimes from "../unit-times";
 import * as Unit from "../unit";
-import { Meter, Second } from "./base-units";
+import { Meter } from "./base-units";
 import { Newton } from "./force";
 import { Kilo, Mega } from "../unit-prefix";
-import { KiloWatt, MegaWatt, GigaWatt, Watt } from "./power";
-import { Hour } from "./duration";
 
 // tslint:disable:variable-name
 
@@ -24,23 +22,7 @@ export const NewtonMeter = UnitTimes.forceByLength(
 );
 export const Kilojoule = Kilo("Kilojoule", Joule);
 export const Megajoule = Mega("Megajoule", Joule);
-export const KiloWattHour = UnitTimes.powerByDuration(
-  "KiloWattHour",
-  KiloWatt,
-  Hour
-);
-export const MegaWattHour = UnitTimes.powerByDuration(
-  "MegaWattHour",
-  MegaWatt,
-  Hour
-);
-export const GigaWattHour = UnitTimes.powerByDuration(
-  "GigaWattHour",
-  GigaWatt,
-  Hour
-);
-export const WattHour = UnitTimes.powerByDuration("WattHour", Watt, Hour);
-export const WattSecond = UnitTimes.powerByDuration("WattSecond", Watt, Second);
+
 export const Btu = Unit.timesNumber("Btu", 52752792631.0 / 50000000.0, Joule);
 export const KiloBtu = Kilo("KiloBtu", Btu);
 export const MegaBtu = Mega("MegaBtu", Btu);
